pyproject.toml
[project]
name = "clinical-research-assistant"
version = "0.1.0"
description = "Multi-agent clinical research assistant using Pydantic-AI, PySide6, and SQLite"
requires-python = ">=3.11"
dependencies = [
    "pydantic-ai[mcp]>=0.2.0",
    "pyside6>=6.6.0",
    "sqlalchemy>=2.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "logfire>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-qt>=4.3.0",
]

[project.scripts]
clinical-assistant = "src.main:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
